requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Hey Murph - openWakeWord trainer
# Python 3.12 + CUDA 13.0
#
# Installazione:
#   python -m venv /venv/main
#   source /venv/main/bin/activate
#   pip install -r requirements.txt
#   python 00_fix_dependencies.py     # applica patch post-install
#   python 01_setup_and_download.py   # clona repo + scarica dataset (~17 GB, una tantum)
#   python 02_training.py
#
# NON inclusi qui (gestiti da 01_setup_and_download.py):
#   - openwakeword         → pip install -e ./openwakeword --no-deps  (dopo git clone)
#   - piper-phonemize-cross / piper-tts  → per piper-sample-generator
#   - datasets / deep-phonemizer         → per il download del dataset
# ─────────────────────────────────────────────────────────────────────────────

--extra-index-url https://download.pytorch.org/whl/cu130

acoustics==0.2.6
ai-edge-litert==2.1.2
annotated-doc==0.0.4
anyio==4.12.1
asttokens==3.0.1
audiomentations==0.43.1
audioread==3.1.0
backports.strenum==1.2.8
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cmudict==1.1.3
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
HyperPyYAML==1.2.3
idna==3.11
importlib_metadata==9.0.0
importlib_resources==6.5.2
iniconfig==2.3.0
joblib==1.5.3
julius==0.2.7
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
lightning-utilities==0.15.3
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.2
mutagen==1.47.0
networkx==3.6.1
numba==0.65.0
numpy==2.0.0
numpy-minmax==0.5.0
numpy-rms==0.6.0
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnx==1.20.1
onnx-ir==0.2.0
onnx2tf==2.4.0
onnxoptimizer==0.4.2
onnxruntime==1.24.3
onnxruntime-gpu==1.24.4
onnxscript==0.6.2
onnxsim-prebuilt==0.4.39.post2
opencv-python==4.11.0.86
packaging
pandas==3.0.2
pillow==12.1.1
piper-phonemize-cross
platformdirs==4.9.4
pluggy==1.6.0
pooch==1.9.0
primePy==1.3
pronouncing==0.2.0
protobuf==4.25.5
psutil==5.9.5
pyarrow==23.0.1
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-stretch==0.3.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==82.0.1
six==1.17.0
sne4onnx==2.0.1
sng4onnx==2.0.1
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.1.0
speexdsp-ns==0.1.2
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
torch==2.10.0+cu130
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.10.0+cu130
torchdata==0.10.0
torchinfo==1.8.0
torchmetrics==1.9.0
torchtext==0.6.0
torchvision==0.25.0+cu130
tqdm==4.67.1
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
webrtcvad-wheels==2.0.14
wheel==0.46.3
zipp==3.23.0
